+++ opal/mca/hwloc/external/configure.m4
@@ -1,6 +1,6 @@
 # -*- shell-script -*-
 #
-# Copyright (c) 2009-2016 Cisco Systems, Inc.  All rights reserved.
+# Copyright (c) 2009-2017 Cisco Systems, Inc.  All rights reserved
 # Copyright (c) 2014-2016 Research Organization for Information Science
 #                         and Technology (RIST). All rights reserved.
 #
@@ -38,7 +38,7 @@ AC_DEFUN([MCA_opal_hwloc_external_POST_CONFIG],[
                               [Version of hwloc])
 
            # Set this variable so that the framework m4 knows what
-           # file to include in opal/mca/hwloc/hwloc.h
+           # file to include in opal/mca/hwloc/hwloc-internal.h
            opal_hwloc_external_basedir=opal/mca/hwloc/external
            opal_hwloc_base_include="$opal_hwloc_external_basedir/external.h"
