+++ opal/mca/hwloc/hwloc1113/hwloc1113.h
@@ -1,6 +1,6 @@
 /* -*- Mode: C; c-basic-offset:4 ; indent-tabs-mode:nil -*- */
 /*
- * Copyright (c) 2011-2013 Cisco Systems, Inc.  All rights reserved.
+ * Copyright (c) 2011-2017 Cisco Systems, Inc.  All rights reserved
  * Copyright (c) 2014-2015 Intel, Inc. All rights reserved.
  * Copyright (c) 2016      Los Alamos National Security, LLC. All rights
  *                         reserved.
@@ -12,7 +12,7 @@
  * $HEADER$
  *
  * When this component is used, this file is included in the rest of
- * the OPAL/ORTE/OMPI code base via opal/mca/hwloc/hwloc.h.  As such,
+ * the OPAL/ORTE/OMPI code base via opal/mca/hwloc/hwloc-internal.h.  As such,
  * this header represents the public interface to this static component.
  */
