+++ orte/mca/rtc/hwloc/rtc_hwloc.c
@@ -1,5 +1,6 @@
 /*
  * Copyright (c) 2014-2015 Intel, Inc. All rights reserved
+ * Copyright (c) 2017      Cisco Systems, Inc.  All rights reserved
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -17,7 +18,7 @@
 #endif  /* HAVE_UNISTD_H */
 #include <string.h>
 
-#include "opal/mca/hwloc/hwloc.h"
+#include "opal/mca/hwloc/hwloc-internal.h"
 #include "opal/util/argv.h"
 #include "opal/util/opal_environ.h"
